+++ src/sentry_sync.c
@@ -496,7 +496,7 @@ sentry__bgworker_setname(sentry_bgworker_t *bgw, const char *thread_name)
 }
 
 #if defined(SENTRY_PLATFORM_UNIX) || defined(SENTRY_PLATFORM_NX)
-#    include "sentry_unix_spinlock.h"
+#    include "sentry_cpu_relax.h"
 
 static sig_atomic_t g_in_signal_handler = 0;
 static sentry_threadid_t g_signal_handling_thread = { 0 };